+++ flask_site/graphing.py
@@ -1,0 +1,1 @@
+""" Module uses plotly to create JSON for graphs usable by the ploltly javascript libs """