+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Welcome to the 'Apex Legends Tracker' changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0]
+### Added
+- This is the final release for the ranked progression chart. 
+To see your ranked progression for the current season you just go to your 'profile' page.
+
 ## [0.8.4]
 More ranked progression work